backend/realtime/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
asyncpg>=0.29,<1.0
aiomqtt>=2.2,<3.0
paho-mqtt>=2.1,<3.0

